Get data sources / collections implemented in the package

This function returns a table of the data sources / collections that can be downloaded and pre-processed using the getData package

getAvailableDataSources()

Value

a data.frame with the data sources / collections dealt by getData, along with details on each data collection

Details

The output data.frame columns are:

"source": The source acronym

"collection": The collection

"...": ...TODO : finish description

Examples

sources<-getAvailableDataSources()
sources

#>                                          Covariate                   Source
#> 1                                      Temperature             MOD11A1.v006
#> 2                                      Temperature             MYD11A1.v006
#> 3                               Vegetation indices             MOD13Q1.v006
#> 4                               Vegetation indices             MYD13Q1.v006
#> 5                               Evapotranspiration             MOD16A2.v006
#> 6                               Evapotranspiration             MYD16A2.v006
#> 7                                         Rainfall             GPM_3IMERGDF
#> 8                                         Rainfall                   TAMSAT
#> 9                        Rainfall - Night of catch             GPM_3IMERGHH
#> 10                           Wind - Night of catch                     ERA5
#> 11 Apparent magnitude of the Moon - Night of catch                  MIRIADE
#> 12               Nighttime lights - Night of catch                VIIRS DNB
#> 13             Elevation and elevation derivatives             SRTMGL1_v003
#> 14                                      Land cover               CGLS-LC100
#> 15                                      Population                     HRSL
#> 16                                      Population WorldPop_100m_Population
#> 17                                   Built-up area                     HRSL
#> 18                                           Roads             OpenSteetMap
#> 19                          Open waters / Wetlands     Global_Surface_Water
#> 20                                                                         
#> 21                                                                         
#> 22                                      Land cover                 REACT_lu
#> 23                                      Population                REACT_pop
#> 24                                   Built-up area            REACT_builtup
#> 25                                        Pedology               REACT_pedo
#>                  Collection TimeSeries
#> 1                     MODIS       TRUE
#> 2                     MODIS       TRUE
#> 3                     MODIS       TRUE
#> 4                     MODIS       TRUE
#> 5                     MODIS       TRUE
#> 6                     MODIS       TRUE
#> 7                       GPM       TRUE
#> 8                    TAMSAT       TRUE
#> 9                       GPM       TRUE
#> 10                     ERA5       TRUE
#> 11                  MIRIADE       TRUE
#> 12                VIIRS DNB       TRUE
#> 13             SRTMGL1_v004      FALSE
#> 14               CGLS-LC101      FALSE
#> 15                     HRSL      FALSE
#> 16 WorldPop_100m_Population      FALSE
#> 17                     HRSL      FALSE
#> 18             OpenSteetMap      FALSE
#> 19     Global_Surface_Water      FALSE
#> 20                                  NA
#> 21                                  NA
#> 22                                  NA
#> 23                                  NA
#> 24                                  NA
#> 25                                  NA
#>                                                                                    Provider
#> 1                                                                                      NASA
#> 2                                                                                      NASA
#> 3                                                                                      NASA
#> 4                                                                                      NASA
#> 5                                                                                      NASA
#> 6                                                                                      NASA
#> 7                                                                                      NASA
#> 8                                                                                    TAMSAT
#> 9                                                                                      NASA
#> 10                                                                                Copenicus
#> 11                                                                                    IMCCE
#> 12                                                                                     NOAA
#> 13                                                                                     NASA
#> 14                                                        Copernicus Global Land Operations
#> 15 Facebook Connectivity Lab and Center for International Earth Science Information Network
#> 16                                                                                 WorldPop
#> 17 Facebook Connectivity Lab and Center for International Earth Science Information Network
#> 18                                                                             OpenSteetMap
#> 19                                                                                      JRC
#> 20                                                                                         
#> 21                                                                                         
#> 22                                                                              REACT / IRD
#> 23                                                                              REACT / IRD
#> 24                                                                              REACT / IRD
#> 25                                                                              REACT / IRD
#>                                                                                                                                     Long.Name
#> 1                                                           MODIS/Terra Land Surface Temperature/Emissivity Daily L3 Global 1km SIN Grid V006
#> 2                                                            MODIS/Aqua Land Surface Temperature/Emissivity Daily L3 Global 1km SIN Grid V006
#> 3                                                                          MODIS/Terra Vegetation Indices 16-Day L3 Global 250m SIN Grid V006
#> 4                                                                           MODIS/Aqua Vegetation Indices 16-Day L3 Global 250m SIN Grid V006
#> 5                                                                       MODIS/Terra Net Evapotranspiration 8-Day L4 Global 500m SIN Grid V006
#> 6                                                                        MODIS/Aqua Net Evapotranspiration 8-Day L4 Global 500m SIN Grid V006
#> 7                                                                          GPM IMERG Final Precipitation L3 1 day 0.1 degree x 0.1 degree V06
#> 8                                                     Tropical Applications of Meteorology using SATellite data and ground-based observations
#> 9                                                                    GPM IMERG Final Precipitation L3 Half Hourly 0.1 degree x 0.1 degree V06
#> 10                                                                                                                                       ERA5
#> 11                                                                            The Virtual Observatory Solar System Object Ephemeris Generator
#> 12                                                                     Visible Infrared Imaging Radiometer Suite (VIIRS) Day/Night Band (DNB)
#> 13                                                                                  NASA Shuttle Radar Topography Mission Global 1 arc second
#> 14                                                                                                           Moderate dynamic land cover 100m
#> 15                                                                                                           High Resolution Settlement Layer
#> 16 Alpha version 2014 estimates of numbers of people per grid square, with national totals adjusted to match UN population division estimates
#> 17                                                                                                           High Resolution Settlement Layer
#> 18                                                                                                                               OpenSteetMap
#> 19                                                                                                                       Global Surface Water
#> 20                                                                                                                                           
#> 21                                                                                                                                           
#> 22                                                                                                                                           
#> 23                                                                                                                                           
#> 24                                                                                                Built-up areas over the REACT project areas
#> 25                                                                                                                                           
#>    Version                                                  DOI
#> 1        6         https://dx.doi.org/10.5067/MODIS/MOD11A1.006
#> 2        6         https://dx.doi.org/10.5067/MODIS/MYD11A1.006
#> 3        6         https://dx.doi.org/10.5067/MODIS/MOD13Q1.006
#> 4        6         https://dx.doi.org/10.5067/MODIS/MYD13Q1.006
#> 5        6         https://dx.doi.org/10.5067/MODIS/MOD16A2.006
#> 6        6         https://dx.doi.org/10.5067/MODIS/MYD16A2.006
#> 7        6           https://doi.org/10.5067/GPM/IMERGDF/DAY/06
#> 8        3                 http://doi.org/10.1038/sdata.2017.63
#> 9        6           https://doi.org/10.5067/GPM/IMERG/3B-HH/06
#> 10       5                                                     
#> 11       2                                                     
#> 12      NA                                                     
#> 13       3 https://dx.doi.org/10.5067/MEASURES/SRTM/SRTMGL1.003
#> 14       2                                                     
#> 15       1                                                     
#> 16       3         10.5258/SOTON/WP00033\n10.5258/SOTON/WP00065
#> 17       1                                                     
#> 18      NA                                                 <NA>
#> 19       1                                  10.1038/nature20584
#> 20      NA                                                     
#> 21      NA                                                     
#> 22      NA                                                     
#> 23      NA                                                     
#> 24      NA                                                     
#> 25      NA                                                     
#>    Spatial.resolution..m. Temporal.resolution Temporal.resolution.unit
#> 1                  1000.0                   1                      day
#> 2                  1000.0                   1                      day
#> 3                   250.0                  16                      day
#> 4                   250.0                  16                      day
#> 5                   500.0                   8                      day
#> 6                   500.0                   8                      day
#> 7                 10000.0                   1                      day
#> 8                  4000.0                   1                      day
#> 9                 10000.0                  30                   minute
#> 10                27000.0                   1                     hour
#> 11                     NA                  NA                     <NA>
#> 12                  450.0                   1                    month
#> 13                   30.0                  NA                     <NA>
#> 14                  100.0                   1                     year
#> 15                   30.0                  NA                     <NA>
#> 16                  100.0                  NA                     <NA>
#> 17                   30.0                  NA                     <NA>
#> 18                     NA                  NA                     <NA>
#> 19                   30.0                   1                     year
#> 20                     NA                  NA                         
#> 21                     NA                  NA                         
#> 22                    1.6                  NA                         
#> 23                    1.6                  NA                         
#> 24                    1.6                  NA                         
#> 25                  100.0                  NA                         
#>                                                                                    Spatial.Coverage
#> 1                                                                                            Global
#> 2                                                                                            Global
#> 3                                                                                            Global
#> 4                                                                                            Global
#> 5                                                                                            Global
#> 6                                                                                            Global
#> 7                                                                                            Global
#> 8                                                                                            Africa
#> 9                                                                                            Global
#> 10                                                                                           Global
#> 11                                                                                           Global
#> 12                                                                                           Global
#> 13                                                                                           Global
#> 14                                                                                           Global
#> 15 Available for some countries : see list here : https://ciesin.columbia.edu/repository/hrsl/#data
#> 16                                                             Africa, Asia, South America, Oceania
#> 17 Available for some countries : see list here : https://ciesin.columbia.edu/repository/hrsl/#data
#> 18                                                                                           Global
#> 19                                                                                           Global
#> 20                                                                                                 
#> 21                                                                                                 
#> 22                                                                                            REACT
#> 23                                                                                            REACT
#> 24                                                                                            REACT
#> 25                                                                                            REACT
#>                                                                                         URL.metadata
#> 1                                                      https://lpdaac.usgs.gov/products/mod11a1v006/
#> 2                                                      https://lpdaac.usgs.gov/products/myd11a1v006/
#> 3                                                      https://lpdaac.usgs.gov/products/mod13q1v006/
#> 4                                                      https://lpdaac.usgs.gov/products/myd13q1v006/
#> 5                                                      https://lpdaac.usgs.gov/products/mod16a2v006/
#> 6                                                      https://lpdaac.usgs.gov/products/myd16a2v006/
#> 7                                        https://disc.gsfc.nasa.gov/datasets/GPM_3IMERGDF_06/summary
#> 8                                                                    https://www.tamsat.org.uk/about
#> 9                                        https://disc.gsfc.nasa.gov/datasets/GPM_3IMERGHH_06/summary
#> 10     https://cds.climate.copernicus.eu/cdsapp#!/dataset/reanalysis-era5-single-levels?tab=overview
#> 11                                                           http://vo.imcce.fr/webservices/miriade/
#> 12                                      https://ngdc.noaa.gov/eog/viirs/download_dnb_composites.html
#> 13                                                     https://lpdaac.usgs.gov/products/srtmgl1v003/
#> 14                                                     https://land.copernicus.eu/global/products/lc
#> 15                                                 https://ciesin.columbia.edu/repository/hrsl/#over
#> 16                         https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0107042
#> 17                                                 https://ciesin.columbia.edu/repository/hrsl/#over
#> 18                                                               https://www.openstreetmap.org/about
#> 19 https://storage.cloud.google.com/global-surface-water/downloads_ancillary/DataUsersGuidev2018.pdf
#> 20                                                                                                  
#> 21                                                                                                  
#> 22                                                                                                  
#> 23                                                                                                  
#> 24                                                                                                  
#> 25                                                                                                  
#>                                                                        URL.programmatic.download
#> 1                            https://opendap.cr.usgs.gov/opendap/hyrax/MOD11A1.006/contents.html
#> 2                            https://opendap.cr.usgs.gov/opendap/hyrax/MYD11A1.006/contents.html
#> 3                            https://opendap.cr.usgs.gov/opendap/hyrax/MOD13Q1.006/contents.html
#> 4                            https://opendap.cr.usgs.gov/opendap/hyrax/MYD13Q1.006/contents.html
#> 5                            https://opendap.cr.usgs.gov/opendap/hyrax/MOD16A2.006/contents.html
#> 6                            https://opendap.cr.usgs.gov/opendap/hyrax/MYD16A2.006/contents.html
#> 7                           https://gpm1.gesdisc.eosdis.nasa.gov/opendap/GPM_L3/GPM_3IMERGDF.06/
#> 8                                                         https://www.tamsat.org.uk/data/archive
#> 9                           https://gpm1.gesdisc.eosdis.nasa.gov/opendap/GPM_L3/GPM_3IMERGHH.06/
#> 10               https://dominicroye.github.io/en/2018/access-to-climate-reanalysis-data-from-r/
#> 11                                               http://vo.imcce.fr/webservices/miriade/?ephemcc
#> 12 https://gis.ngdc.noaa.gov/arcgis/rest/services/NPP_VIIRS_DNB/Monthly_AvgRadiance/ImageServer/
#> 13                                  https://e4ftl01.cr.usgs.gov/MEASURES/SRTMGL1.003/2000.02.11/
#> 14                                                                                          <NA>
#> 15                                                                                          <NA>
#> 16                                                                                          <NA>
#> 17                                                                                          <NA>
#> 18                                                                                              
#> 19                                                                                          <NA>
#> 20                                                                                              
#> 21                                                                                              
#> 22                                                                                              
#> 23                                                                                              
#> 24                                                                                              
#> 25                                                                                              
#>                                                                                    URL.manual.download
#> 1                                        https://search.earthdata.nasa.gov/search?q=MOD11A1&ok=MOD11A1
#> 2                                        https://search.earthdata.nasa.gov/search?q=MYD11A1&ok=MYD11A1
#> 3                                        https://search.earthdata.nasa.gov/search?q=MOD13Q1&ok=MOD13Q1
#> 4                                        https://search.earthdata.nasa.gov/search?q=MYD13Q1&ok=MYD13Q1
#> 5                                        https://search.earthdata.nasa.gov/search?q=MOD16A2&ok=MOD16A2
#> 6                                        https://search.earthdata.nasa.gov/search?q=MYD16A2&ok=MYD16A2
#> 7                                           https://search.earthdata.nasa.gov/search?q=GPM_3IMERGDF_06
#> 8                                                         https://www.tamsat.org.uk/data/rfe/index.cgi
#> 9                                                                                                     
#> 10           https://cds.climate.copernicus.eu/cdsapp#!/dataset/reanalysis-era5-single-levels?tab=form
#> 11                                                       http://vo.imcce.fr/webservices/miriade/?forms
#> 12                                        https://ngdc.noaa.gov/eog/viirs/download_dnb_composites.html
#> 13 https://search.earthdata.nasa.gov/search/collection-details?p=C1000000240-LPDAAC_ECS&q=SRTM&ok=SRTM
#> 14                                                                   https://lcviewer.vito.be/download
#> 15                                                   https://ciesin.columbia.edu/repository/hrsl/#data
#> 16                                                      https://www.worldpop.org/geodata/listing?id=16
#> 17                                                   https://ciesin.columbia.edu/repository/hrsl/#data
#> 18                                                                      https://www.openstreetmap.org/
#> 19                                                   https://global-surface-water.appspot.com/download
#> 20                                                                                                    
#> 21                                                                                                    
#> 22                                                                                                    
#> 23                                                                                                    
#> 24                                                                                                    
#> 25                                                                                                    
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                           Citation
#> 1                                                                                                                                                                                                                                                                                                                                      Wan, Z., Hook, S., Hulley, G. (2015). MOD11A1 MODIS/Terra Land Surface Temperature/Emissivity Daily L3 Global 1km SIN Grid V006 [Data set]. NASA EOSDIS Land Processes DAAC. doi: 10.5067/MODIS/MOD11A1.006
#> 2                                                                                                                                                                                                                                                                                                                                       Wan, Z., Hook, S., Hulley, G. (2015). MYD11A1 MODIS/Aqua Land Surface Temperature/Emissivity Daily L3 Global 1km SIN Grid V006 [Data set]. NASA EOSDIS Land Processes DAAC. doi: 10.5067/MODIS/MYD11A1.006
#> 3                                                                                                                                                                                                                                                                                                                                                                         Didan, K. (2015). MOD13Q1 MODIS/Terra Vegetation Indices 16-Day L3 Global 250m SIN Grid V006 [Data set]. NASA EOSDIS Land Processes DAAC. doi: 10.5067/MODIS/MOD13Q1.006
#> 4                                                                                                                                                                                                                                                                                                                                                                          Didan, K. (2015). MYD13Q1 MODIS/Aqua Vegetation Indices 16-Day L3 Global 250m SIN Grid V006 [Data set]. NASA EOSDIS Land Processes DAAC. doi: 10.5067/MODIS/MYD13Q1.006
#> 5                                                                                                                                                                                                                                                                                                                                                  Running, S., Mu, Q., Zhao, M. (2017). MOD16A2 MODIS/Terra Net Evapotranspiration 8-Day L4 Global 500m SIN Grid V006 [Data set]. NASA EOSDIS Land Processes DAAC. doi: 10.5067/MODIS/MOD16A2.006
#> 6                                                                                                                                                                                                                                                                                                                                                   Running, S., Mu, Q., Zhao, M. (2017). MYD16A2 MODIS/Aqua Net Evapotranspiration 8-Day L4 Global 500m SIN Grid V006 [Data set]. NASA EOSDIS Land Processes DAAC. doi: 10.5067/MODIS/MYD16A2.006
#> 7                                                                                                                                                                                                                      Huffman, G.J., E.F. Stocker, D.T. Bolvin, E.J. Nelkin, Jackson Tan (2019), GPM IMERG Final Precipitation L3 1 day 0.1 degree x 0.1 degree V06, Edited by Andrey Savtchenko, Greenbelt, MD, Goddard Earth Sciences Data and Information Services Center (GES DISC), Accessed: [Data Access Date], 10.5067/GPM/IMERGDF/DAY/06
#> 8  Tarnavsky, E., D. Grimes, R. Maidment, E. Black, R. Allan, M. Stringer, R. Chadwick, F. Kayitakire (2014) Extension of the TAMSAT Satellite-based Rainfall Monitoring over Africa and from 1983 to present Journal of Applied Meteorology and Climate DOI 10.1175/JAMC-D-14-0016.1\nMaidment, R., D. Grimes, R.P.Allan, E. Tarnavsky, M. Stringer, T. Hewison, R. Roebeling and E. Black (2014) The 30 year TAMSAT African Rainfall Climatology And Time series (TARCAT) data set Journal of Geophysical Research DOI: 10.1002/2014JD021927\n\n
#> 9                                                                                                                                                                                                                                             Huffman, G.J., E.F. Stocker, D.T. Bolvin, E.J. Nelkin, Jackson Tan (2019), GPM IMERG Final Precipitation L3 Half Hourly 0.1 degree x 0.1 degree V06, Greenbelt, MD, Goddard Earth Sciences Data and Information Services Center (GES DISC), Accessed: [Data Access Date], 10.5067/GPM/IMERG/3B-HH/06
#> 10                                                                                                                                                                                                                                                                                      Copernicus Climate Change Service (C3S) (2017): ERA5: Fifth generation of ECMWF atmospheric reanalyses of the global climate . Copernicus Climate Change Service Climate Data Store (CDS), date of access. https://cds.climate.copernicus.eu/cdsapp#!/home
#> 11                                                                                                                                                                                                                                                                                                                                                                                                                                                                                           This research has made use of IMCCE's Miriade VO tool
#> 12                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                
#> 13                                                                                                                                                                                                                                                                                                                                                                                  NASA JPL (2013). NASA Shuttle Radar Topography Mission Global 1 arc second [Data set]. NASA EOSDIS Land Processes DAAC. doi: 10.5067/MEaSUREs/SRTM/SRTMGL1.003
#> 14                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                
#> 15                                                                                                                                                                                                                                                                                                   Facebook Connectivity Lab and Center for International Earth Science Information Network - CIESIN - Columbia University. 2016. High Resolution Settlement Layer (HRSL). Source imagery for HRSL © 2016 DigitalGlobe. Accessed DAY MONTH YEAR.
#> 16                                                                                                                                                                                            WorldPop (www.worldpop.org - School of Geography and Environmental Science, University of Southampton). 2013. Burkina Faso 100m Population. Alpha version 2010 and 2014 estimates of numbers of people per grid square, with national totals adjusted to match UN population division estimates (http://esa.un.org/wpp/). DOI: 10.5258/SOTON/WP00033
#> 17                                                                                                                                                                                                                                                                                                   Facebook Connectivity Lab and Center for International Earth Science Information Network - CIESIN - Columbia University. 2016. High Resolution Settlement Layer (HRSL). Source imagery for HRSL © 2016 DigitalGlobe. Accessed DAY MONTH YEAR.
#> 18                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                
#> 19                                                                                                                                                                                                                                                                                                                                            Jean-Francois Pekel, Andrew Cottam, Noel Gorelick, Alan S. Belward, High-resolution mapping of global surface water and its long-term changes. Nature 540, 418-422 (2016). (doi:10.1038/nature20584)
#> 20                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                
#> 21                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                
#> 22                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                
#> 23                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                
#> 24                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                
#> 25                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                
#>    Rfunction_getData Rfunction_processData
#> 1      getData_modis     prepareData_modis
#> 2      getData_modis     prepareData_modis
#> 3      getData_modis     prepareData_modis
#> 4      getData_modis     prepareData_modis
#> 5      getData_modis     prepareData_modis
#> 6      getData_modis     prepareData_modis
#> 7        getData_gpm       prepareData_gpm
#> 8     getData_tamsat    prepareData_tamsat
#> 9        getData_gpm       prepareData_gpm
#> 10      getData_era5                      
#> 11   getData_miriade                      
#> 12  getData_viirsdnb                      
#> 13      getData_srtm                      
#> 14      getData_cgls                      
#> 15      getData_hrsl                      
#> 16  getData_worldpop                      
#> 17      getData_hrsl                      
#> 18                                        
#> 19       getData_gsw                      
#> 20                                        
#> 21                                        
#> 22     getData_react                      
#> 23     getData_react                      
#> 24     getData_react                      
#> 25     getData_react