Use esri2sf to get data from an ArcGIS FeatureServer or MapServer for a location

Wraps the esri2sf::esri2sf and esri2sf::esri2df function to download an ArcGIS FeatureServer or MapServer. Supports spatial filtering with bounding box based on location and filtering by location name (if location name column is provided). As of July 2022, this package suggests the elipousson/esri2sf fork using httr2 but I expect to see this fork merged back into the main package in the future.

Usage

get_esri_data(
  location = NULL,
  dist = getOption("getdata.dist"),
  diag_ratio = getOption("getdata.diag_ratio"),
  unit = getOption("getdata.unit"),
  asp = getOption("getdata.asp"),
  crs = getOption("getdata.crs", 3857),
  url,
  where = NULL,
  name = NULL,
  name_col = NULL,
  coords = NULL,
  from_crs = getOption("getdata.crs", 4326),
  clean_names = TRUE,
  progress = TRUE,
  ...
)

get_esri_layers(location = NULL, layers, url = NULL, nm = NULL, ...)

get_esri_metadata(url, meta = "name", clean_names = TRUE)

Arguments

location

sf, sfc, or bbox object (or other object convertible with as_bbox(). Optional.

dist

buffer distance in units. Optional.

diag_ratio

ratio of diagonal distance of area's bounding box used as buffer distance. e.g. if the diagonal distance is 3000 meters and the "diag_ratio = 0.1" a 300 meter will be used. Ignored when dist is provided.

unit

Units for buffer. Supported options include "meter", "foot", "kilometer", and "mile", "nautical mile" Common abbreviations (e.g. "km" instead of "kilometer") are also supported. Distance in units is converted to units matching GDAL units for x; defaults to "meter"

asp

Aspect ratio of width to height as a numeric value (e.g. 0.33) or character (e.g. "1:3"). If numeric, get_asp() returns the same value without modification.

crs

Cordinate reference system to return, Default: 4326 for sf_to_df and NULL for df_to_sf.

url

FeatureServer or MapServer url to retrieve data from. Passed to url parameter of esri2sf::esri2sf or esri2sf::esri2df functions. For get_esri_layers, the optional url must be a service url which is the base url for one or more layer urls.

where

where query string passed to esri2sf, Default: NULL

name, name_col

Name value and name column found in the ArcGIS FeatureServer or MapServer data.

coords

Coordinate columns for input dataframe or output sf object (if geometry is 'centroid' or 'point') Default: c("lon", "lat").

from_crs

For df_to_sf, coordinate reference system used by coordinates or well known text in data frame.

clean_names

If TRUE, pass data frame to janitor::clean_names; defaults to TRUE.

progress

Show progress bar from cli::cli_progress_along() if TRUE. Default FALSE.

...

Arguments passed on to esri2sf::esri2sf

outFields

vector of fields you want to include. default is NULL for all fields.

geometry

An sf or bbox object. Currently, sf objects with a single POINT feature are supported. All other sf objects are converted to bbox objects.

bbox

bbox class object from sf::st_bbox() or a simple feature object that can be converted to a bounding box.

token

string for authentication token. defaults to NULL.

geomType

string specifying the layer geometry ('esriGeometryPolygon' or 'esriGeometryPoint' or 'esriGeometryPolyline' - if NULL, will try to be inferred from the server)

spatialRel

Spatial relationship applied to the input geometry when performing the query; defaults to NULL (equivalent to "esriSpatialRelIntersects"). Additional supported options include "esriSpatialRelContains", "esriSpatialRelCrosses", "esriSpatialRelEnvelopeIntersects", "esriSpatialRelIndexIntersects", "esriSpatialRelOverlaps", "esriSpatialRelTouches", "esriSpatialRelWithin"

replaceDomainInfo

If TRUE, add domain information to the return data frame. Default FALSE.

layers

Either a vector with URLs, a named list of urls, or a numeric vector.

nm

Name or vector of names to add to the layers; defaults to NULL.

meta

Name of metadata list value to return from esri2sf::esrimeta.

clean

If TRUE, use janitor::make_clean_names on the returned metadata value (typically used for name values).

See also