Format data frames and simple features with address data

getdata has two helpers for working with address data:

bind_address_col() bind a provided value for city, county, and state to a data frame (to supplement address data with consistent values for these variables). This function is useful for converting partial street addresses with a consistent values for state, county, or city into full addresses

bind_block_col() requires a data frame with columns named "bldg_num", "street_dir_prefix", "street_name", and "street_type" and binds derived values for whether a building is on the even or odd side of a block and create a block segment and a block face (including the even/odd identifier).

Usage

bind_block_col(
  x,
  bldg_num = "bldg_num",
  street_dir_prefix = "street_dir_prefix",
  street_name = "street_name",
  street_suffix = "street_type",
  replace_suffix = FALSE,
  block_col = NULL,
  .after = street_suffix,
  case = NULL
)

bind_address_col(
  x,
  ...,
  case = NULL,
  street = "street_address",
  address = "address",
  address_cols = list(city = "city", county = "county", state = "state"),
  .after = NULL
)

Arguments

x

A data frame to bind the address columns to.

bldg_num, street_dir_prefix, street_name, street_suffix

Column names to use for address information required to generate a block name and number.

replace_suffix

If TRUE, replace values in street_suffix column with abbreviations from street_suffixes.

block_col

String to use as prefix for block identifier columns and separator between block number and street. Set to "block" when NULL (default). If length 2 (e.g. c("blk", "block")), the second value is used as the block separator and the first as the column identifier prefix.

.after

passed to dplyr::mutate() defaults to street_suffix for bind_block_col() and "address" for bind_address_col().

case

Case to use for text in new columns or in modified values. Options include "lower", "upper", "title", or "sentence". Defaults to NULL which leaves the case as is.

...

Additional parameters passed to dplyr::mutate() intended for use in filling missing values, e.g. state = "MD" to add a missing state column.

street

Street address column name, (e.g. 100 Holliday St) Default: 'street_address'.

address

Full address column name. If city and state or city and county are provided, a combined address column is added to the data.frame using the street address column. If both county and city are provided, county is ignored.

address_cols

Named list specifying the additional column names to use for city, county, and state data.