Format data frames and simple features using common approaches

This function can apply the following common data cleaning tasks:

Usage

format_data(
  x,
  var_names = NULL,
  clean_names = TRUE,
  replace_na_with = NULL,
  replace_with_na = NULL,
  replace_empty_char_with_na = TRUE,
  fix_date = FALSE,
  label = FALSE,
  format_sf = FALSE,
  ...
)

label_with_xwalk(x, xwalk)

rename_with_xwalk(x, xwalk = NULL, label = FALSE)

fix_date(x)

bind_address_col(x, city = NULL, county = NULL, state = NULL)

bind_block_col(
  x,
  bldg_num = "bldg_num",
  street_dir_prefix = "street_dir_prefix",
  street_name = "street_name",
  street_suffix = "street_type"
)

Arguments

x

A tibble or data frame object

var_names

A named list following the format, list("New var name" = old_var_name), or a two column data frame with the first column being the new variable names and the second column being the old variable names; defaults to NULL.

clean_names

If TRUE, pass data frame to janitor::clean_names; defaults to TRUE.

replace_na_with

A named list to pass to tidyr::replace_na; defaults to NULL.

replace_with_na

A named list to pass to naniar::replace_with_na; defaults to NULL.

replace_empty_char_with_na

If TRUE, replace "" with NA using naniar::replace_with_na_if, Default: TRUE

fix_date

If TRUE, fix UNIX dates (common issue with dates from FeatureServer and MapServer sources) , Default: TRUE

...

Additional parameters passed to format_sf_data

xwalk

a data frame with two columns using the first column as name and the second column as value; or a named list. The existing names of x must be the values and the new names must be the names.

city, county, state

City, county, and state to bind to data frame or sf object.

Value

The input data frame or simple feature object with formatting functions applied.

Details

Applies stringr::str_squish and stringr::str_trim to all character columns (str_trim_squish)

Optionally replaces all character values of "" with NA values

Optionally corrects UNIX formatted dates with 1970-01-01 origins

Optionally renames variables by passing a named list of variables

Bind columns to address data:

bind_address_col bind a provided value for city, county, and state to a data frame (to supplement address data with consistent values for these variables)

bind_block_col requires a data frame with columns named "bldg_num", "street_dir_prefix", "street_name", and "street_type" and binds derived values for whether a building is on the even or odd side of a block and create a block number (street segment), and block face (street segment side) identifier.