Get Airtable data

Get Airtable data

Usage

get_airtable_data(
  base,
  table,
  record = NULL,
  fields = NULL,
  filter = NULL,
  sort = NULL,
  desc = FALSE,
  view = NULL,
  max_records = NULL,
  per_page = NULL,
  cell_format = "json",
  tz = NULL,
  locale = NULL,
  fields_by_id = FALSE,
  offset = NULL,
  list = "records",
  geometry = TRUE,
  location = NULL,
  dist = getOption("getdata.dist"),
  diag_ratio = getOption("getdata.diag_ratio"),
  unit = getOption("getdata.unit", "meter"),
  asp = getOption("getdata.asp"),
  crs = getOption("getdata.crs", 3857),
  coords = getOption("getdata.coords", c("lon", "lat")),
  from_crs = getOption("getdata.from_crs", 4326),
  remove_coords = TRUE,
  address = getOption("getdata.address", "address"),
  geo = FALSE,
  clean_names = TRUE,
  label = TRUE,
  token = NULL,
  type = "AIRTABLE_API_KEY"
)

Arguments

base

Airtable base identifier

table

Airtable table name or identifier

record

Airtable record identifier, Default: NULL

fields

Fields to return from Airtable base, Default: NULL

filter

Filter (not currently supported), Default: NULL

sort

Field to sort by, Default: NULL

desc

If TRUE, sort in descending order, Default: FALSE

view

Airtable view name(?) or identifier, Default: NULL

max_records

Max records to return, Default: NULL

per_page

Max records to return per page, Default: NULL

cell_format

Cell format, Default: 'json'

tz

Time zone, Default: NULL

locale

Locale, Default: NULL

fields_by_id

If TRUE, return fields by id, Default: FALSE

offset

Offset parameter, Default: NULL

list

Data type to return, Default: 'records'

geometry

Type of geometry to include in data frame. options include "drop", "wkt", "centroid", "point", Default: 'centroid'.

location

sf object. If multiple areas are provided, they are unioned into a single sf object using sf::st_union

dist

buffer distance in units. Optional.

diag_ratio

ratio of diagonal distance of area's bounding box used as buffer distance. e.g. if the diagonal distance is 3000 meters and the "diag_ratio = 0.1" a 300 meter will be used. Ignored when dist is provided.

unit

Units for buffer. Supported options include "meter", "foot", "kilometer", and "mile", "nautical mile" Common abbreviations (e.g. "km" instead of "kilometer") are also supported. Distance in units is converted to units matching GDAL units for x; defaults to "meter"

asp

Aspect ratio of width to height as a numeric value (e.g. 0.33) or character (e.g. "1:3"). If numeric, get_asp() returns the same value without modification.

crs

Cordinate reference system to return, Default: 4326 for sf_to_df and NULL for df_to_sf.

coords

Coordinate columns for input dataframe or output sf object (if geometry is 'centroid' or 'point') Default: c("lon", "lat").

from_crs

For df_to_sf, coordinate reference system used by coordinates or well known text in data frame.

remove_coords

For df_to_sf, if TRUE, remove the coordinate columns after converting a data frame to simple feature object; defaults to FALSE.

address

Address column name passed to tidygeocoder::geocode or tidygeocoder::geo

geo

If TRUE, use address_to_sf to geocode address column; defaults to FALSE.

clean_names

If TRUE, clean names provided to nm or created based on value of col using janitor::clean_names. If FALSE, use names as provided.

label

label is optionally used by map_location_data to name the data objects in the list returned by the function.

token, type

API token and type, token defaults to NULL and type to "AIRTABLE_API_KEY" (same as get_access_token(type = "AIRTABLE_API_KEY"))